Undo model moves Fixs required attributes on delegate model Removes wrong startBlock attribute on delegate model Adds reward ref on delegateModel
